implement kdc todo
