+++ kedro/io/kedro_data_catalog.py
@@ -10,22 +10,21 @@ Expect possible breaking changes while using it.
 
 from __future__ import annotations
 
-import difflib
 import logging
 import re
 from typing import Any, Iterator, List  # noqa: UP035
 
-from kedro.io.catalog_config_resolver import CatalogConfigResolver, Patterns
+from kedro.io.cached_dataset import CachedDataset
+from kedro.io.catalog_config_resolver import CatalogConfigResolver
 from kedro.io.core import (
     TYPE_KEY,
     AbstractDataset,
     AbstractVersionedDataset,
     CatalogProtocol,
-    DatasetAlreadyExistsError,
     DatasetError,
     DatasetNotFoundError,
     Version,
-    _validate_versions,
+    VersionAlreadyExistsError,
     generate_timestamp,
     parse_dataset_definition,
 )
@@ -100,10 +99,9 @@ class KedroDataCatalog(CatalogProtocol):
             >>> catalog = KedroDataCatalog(datasets={"cars": cars})
         """
         self._config_resolver = config_resolver or CatalogConfigResolver()
-        # TODO: rename back to _datasets when removing old catalog
-        self.__datasets: dict[str, AbstractDataset] = datasets or {}
+        self._datasets: dict[str, AbstractDataset] = datasets or {}
         self._lazy_datasets: dict[str, _LazyDataset] = {}
-        self._load_versions, self._save_version = _validate_versions(
+        self._load_versions, self._save_version = self._validate_versions(
             datasets, load_versions or {}, save_version
         )
 
@@ -117,37 +115,16 @@ class KedroDataCatalog(CatalogProtocol):
             self[ds_name] = data  # type: ignore[has-type]
 
     @property
-    def datasets(self) -> dict[str, Any]:
-        # TODO: remove when removing old catalog
-        return self._lazy_datasets | self.__datasets
-
-    @datasets.setter
-    def datasets(self, value: Any) -> None:
-        # TODO: remove when removing old catalog
-        raise AttributeError(
-            "Operation not allowed. Please use KedroDataCatalog.add() instead."
-        )
-
-    def __getattribute__(self, key: str) -> Any:
-        # Needed for compatability with old catalog interface since now we
-        # differ _datasets and _lazy_datasets
-        # TODO: remove when removing old catalog
-        if key == "_datasets":
-            return self.datasets
-        else:
-            return super().__getattribute__(key)
-
-    @property
     def config_resolver(self) -> CatalogConfigResolver:
         return self._config_resolver
 
     def __repr__(self) -> str:
-        return repr(self._lazy_datasets | self.__datasets)
+        return repr(self._lazy_datasets | self._datasets)
 
     def __contains__(self, dataset_name: str) -> bool:
         """Check if an item is in the catalog as a materialised dataset or pattern."""
         return (
-            dataset_name in self.__datasets
+            dataset_name in self._datasets
             or dataset_name in self._lazy_datasets
             or self._config_resolver.match_pattern(dataset_name) is not None
         )
@@ -155,18 +132,18 @@ class KedroDataCatalog(CatalogProtocol):
     def __eq__(self, other) -> bool:  # type: ignore[no-untyped-def]
         """Compares two catalogs based on materialised datasets and datasets patterns."""
         return (
-            self.__datasets,
+            self._datasets,
             self._lazy_datasets,
             self._config_resolver.list_patterns(),
         ) == (
-            other.__datasets,
+            other._datasets,
             other._lazy_datasets,
             other.config_resolver.list_patterns(),
         )
 
     def keys(self) -> List[str]:  # noqa: UP006
         """List all dataset names registered in the catalog."""
-        return list(self._lazy_datasets.keys()) + list(self.__datasets.keys())
+        return list(self._lazy_datasets.keys()) + list(self._datasets.keys())
 
     def values(self) -> List[AbstractDataset]:  # noqa: UP006
         """List all datasets registered in the catalog."""
@@ -195,7 +172,7 @@ class KedroDataCatalog(CatalogProtocol):
             DatasetNotFoundError: When a dataset with the given name
                 is not in the collection and does not match patterns.
         """
-        return self.get_dataset(ds_name)
+        return self.get(ds_name)
 
     def __setitem__(self, key: str, value: Any) -> None:
         """Add dataset to the ``KedroDataCatalog`` using the given key as a datsets name
@@ -230,24 +207,27 @@ class KedroDataCatalog(CatalogProtocol):
             >>>
             >>> assert catalog.load("data_csv_dataset").equals(df)
         """
-        if key in self.__datasets:
+        if key in self._datasets:
             self._logger.warning("Replacing dataset '%s'", key)
         if isinstance(value, AbstractDataset):
-            self._load_versions, self._save_version = _validate_versions(
+            self._load_versions, self._save_version = self._validate_versions(
                 {key: value}, self._load_versions, self._save_version
             )
-            self.__datasets[key] = value
+            self._datasets[key] = value
         elif isinstance(value, _LazyDataset):
             self._lazy_datasets[key] = value
         else:
             self._logger.info(f"Adding input data as a MemoryDataset - {key}")
-            self.__datasets[key] = MemoryDataset(data=value)  # type: ignore[abstract]
+            self._datasets[key] = MemoryDataset(data=value)  # type: ignore[abstract]
 
     def __len__(self) -> int:
         return len(self.keys())
 
     def get(
-        self, key: str, default: AbstractDataset | None = None
+        self,
+        key: str,
+        version: Version | None = None,
+        default: AbstractDataset | None = None,
     ) -> AbstractDataset | None:
         """Get a dataset by name from an internal collection of datasets.
 
@@ -256,13 +236,14 @@ class KedroDataCatalog(CatalogProtocol):
 
         Args:
             key: A dataset name.
+            version: Optional argument to get a specific version of the dataset.
             default: Optional argument for default dataset to return in case
                 requested dataset not in the catalog.
 
         Returns:
             An instance of AbstractDataset.
         """
-        if key not in self.__datasets and key not in self._lazy_datasets:
+        if key not in self._datasets and key not in self._lazy_datasets:
             ds_config = self._config_resolver.resolve_pattern(key)
             if ds_config:
                 self._add_from_config(key, ds_config)
@@ -271,7 +252,12 @@ class KedroDataCatalog(CatalogProtocol):
         if lazy_dataset:
             self[key] = lazy_dataset.materialize()
 
-        dataset = self.__datasets.get(key, None)
+        dataset = self._datasets.get(key, None)
+
+        if version and isinstance(dataset, AbstractVersionedDataset):
+            # we only want to return a similar-looking dataset,
+            # not modify the one stored in the current catalog
+            dataset = dataset._copy(_version=version)
 
         return dataset or default
 
@@ -437,7 +423,7 @@ class KedroDataCatalog(CatalogProtocol):
             credentials.update(unresolved_credentials)
             load_versions[ds_name] = self._load_versions.get(ds_name, None)
 
-        for ds_name, ds in self.__datasets.items():  # type: ignore[assignment]
+        for ds_name, ds in self._datasets.items():  # type: ignore[assignment]
             if _is_memory_dataset(ds):  # type: ignore[arg-type]
                 continue
             resolved_config = ds.to_config()  # type: ignore[attr-defined]
@@ -486,71 +472,6 @@ class KedroDataCatalog(CatalogProtocol):
         )
 
         self.add(ds_name, ds)
-
-    def get_dataset(
-        self, ds_name: str, version: Version | None = None, suggest: bool = True
-    ) -> AbstractDataset:
-        # TODO: remove when removing old catalog
-        """Get a dataset by name from an internal collection of datasets.
-
-        If a dataset is not in the collection but matches any pattern
-        it is instantiated and added to the collection first, then returned.
-
-        Args:
-            ds_name: A dataset name.
-            version: Optional argument for concrete dataset version to be loaded.
-                Works only with versioned datasets.
-            suggest: Optional argument whether to suggest fuzzy-matching datasets' names
-                in the DatasetNotFoundError message.
-
-        Returns:
-            An instance of AbstractDataset.
-
-        Raises:
-            DatasetNotFoundError: When a dataset with the given name
-                is not in the collection and do not match patterns.
-        """
-        dataset = self.get(ds_name)
-
-        if dataset is None:
-            error_msg = f"Dataset '{ds_name}' not found in the catalog"
-            # Flag to turn on/off fuzzy-matching which can be time consuming and
-            # slow down plugins like `kedro-viz`
-            if suggest:
-                matches = difflib.get_close_matches(ds_name, self.keys())
-                if matches:
-                    suggestions = ", ".join(matches)
-                    error_msg += f" - did you mean one of these instead: {suggestions}"
-            raise DatasetNotFoundError(error_msg)
-
-        if version and isinstance(dataset, AbstractVersionedDataset):
-            # we only want to return a similar-looking dataset,
-            # not modify the one stored in the current catalog
-            dataset = dataset._copy(_version=version)
-
-        return dataset
-
-    def _get_dataset(
-        self, dataset_name: str, version: Version | None = None, suggest: bool = True
-    ) -> AbstractDataset:
-        # TODO: remove when removing old catalog
-        return self.get_dataset(dataset_name, version, suggest)
-
-    def add(
-        self,
-        ds_name: str,
-        dataset: AbstractDataset | _LazyDataset,
-        replace: bool = False,
-    ) -> None:
-        # TODO: remove when removing old catalog
-        """Adds a new ``AbstractDataset`` object to the ``KedroDataCatalog``."""
-        if (
-            ds_name in self.__datasets or ds_name in self._lazy_datasets
-        ) and not replace:
-            raise DatasetAlreadyExistsError(
-                f"Dataset '{ds_name}' has already been registered"
-            )
-        self.__setitem__(ds_name, dataset)
 
     def filter(
         self,
@@ -618,7 +539,7 @@ class KedroDataCatalog(CatalogProtocol):
                 if ds_name in self._lazy_datasets:
                     str_type = str(self._lazy_datasets[ds_name])
                 else:
-                    class_type = type(self.__datasets[ds_name])
+                    class_type = type(self._datasets[ds_name])
                     str_type = f"{class_type.__module__}.{class_type.__qualname__}"
                 # Match against type_regex and apply by_type filtering
                 if (not type_regex or re.search(type_regex, str_type)) and (
@@ -630,61 +551,11 @@ class KedroDataCatalog(CatalogProtocol):
 
         return filtered
 
-    def list(
-        self, regex_search: str | None = None, regex_flags: int | re.RegexFlag = 0
-    ) -> List[str]:  # noqa: UP006
-        # TODO: remove when removing old catalog
-        """List all dataset names registered in the catalog, optionally filtered by a regex pattern.
-
-        If a regex pattern is provided, only dataset names matching the pattern will be returned.
-        This method supports optional regex flags for customization
-
-        Args:
-            regex_search: Optional regular expression to filter dataset names.
-            regex_flags: Optional regex flags.
-        Returns:
-            A list of dataset names that match the `regex_search` criteria. If no pattern is
-                provided, all dataset names are returned.
-
-        Raises:
-            SyntaxError: If the provided regex pattern is invalid.
-
-        Example:
-        ::
-
-            >>> catalog = KedroDataCatalog()
-            >>> # get datasets where the substring 'raw' is present
-            >>> raw_data = catalog.list(regex_search='raw')
-            >>> # get datasets which start with 'prm' or 'feat'
-            >>> feat_eng_data = catalog.list(regex_search='^(prm|feat)')
-            >>> # get datasets which end with 'time_series'
-            >>> models = catalog.list(regex_search='.+time_series$')
-        """
-        if regex_search is None:
-            return self.keys()
-
-        if regex_search == "":
-            self._logger.warning("The empty string will not match any datasets")
-            return []
-
-        if not regex_flags:
-            regex_flags = re.IGNORECASE
-
-        try:
-            pattern = re.compile(regex_search, flags=regex_flags)
-        except re.error as exc:
-            raise SyntaxError(
-                f"Invalid regular expression provided: '{regex_search}'"
-            ) from exc
-
-        return [ds_name for ds_name in self.__iter__() if pattern.search(ds_name)]
-
-    def save(self, name: str, data: Any) -> None:
-        # TODO: rename input argument when breaking change: name -> ds_name
+    def save(self, ds_name: str, data: Any) -> None:
         """Save data to a registered dataset.
 
         Args:
-            name: A dataset to be saved to.
+            ds_name: The name of the dataset to be saved.
             data: A data object to be saved as configured in the registered
                 dataset.
 
@@ -710,24 +581,22 @@ class KedroDataCatalog(CatalogProtocol):
             >>>                    'col3': [5, 6]})
             >>> catalog.save("cars", df)
         """
-        dataset = self.get_dataset(name)
+        dataset = self.get(ds_name)
 
         self._logger.info(
             "Saving data to %s (%s)...",
-            _format_rich(name, "dark_orange") if self._use_rich_markup else name,
+            _format_rich(ds_name, "dark_orange") if self._use_rich_markup else ds_name,
             type(dataset).__name__,
             extra={"markup": True},
         )
 
         dataset.save(data)
 
-    def load(self, name: str, version: str | None = None) -> Any:
-        # TODO: rename input argument when breaking change: name -> ds_name
-        # TODO: remove version from input arguments when breaking change
+    def load(self, ds_name: str, version: str | None = None) -> Any:
         """Loads a registered dataset.
 
         Args:
-            name: A dataset to be loaded.
+            ds_name: The name of the dataset to be loaded.
             version: Optional argument for concrete data version to be loaded.
                 Works only with versioned datasets.
 
@@ -752,11 +621,11 @@ class KedroDataCatalog(CatalogProtocol):
             >>> df = catalog.load("cars")
         """
         load_version = Version(version, None) if version else None
-        dataset = self.get_dataset(name, version=load_version)
+        dataset = self.get(ds_name, version=load_version)
 
         self._logger.info(
             "Loading data from %s (%s)...",
-            _format_rich(name, "dark_orange") if self._use_rich_markup else name,
+            _format_rich(ds_name, "dark_orange") if self._use_rich_markup else ds_name,
             type(dataset).__name__,
             extra={"markup": True},
         )
@@ -771,7 +640,7 @@ class KedroDataCatalog(CatalogProtocol):
             DatasetNotFoundError: When a dataset with the given name
                 has not yet been registered.
         """
-        dataset = self.get_dataset(name)
+        dataset = self.get(name)
         dataset.release()
 
     def confirm(self, name: str) -> None:
@@ -782,32 +651,12 @@ class KedroDataCatalog(CatalogProtocol):
             DatasetError: When the dataset does not have `confirm` method.
         """
         self._logger.info("Confirming dataset '%s'", name)
-        dataset = self.get_dataset(name)
+        dataset = self.get(name)
 
         if hasattr(dataset, "confirm"):
             dataset.confirm()
         else:
             raise DatasetError(f"Dataset '{name}' does not have 'confirm' method")
-
-    def add_feed_dict(self, feed_dict: dict[str, Any], replace: bool = False) -> None:
-        # TODO: remove when removing old catalog
-        # This method was simplified to add memory datasets only, since
-        # adding AbstractDataset can be done via add() method
-        for ds_name, ds_data in feed_dict.items():
-            self.add(ds_name, MemoryDataset(data=ds_data), replace)  # type: ignore[abstract]
-
-    def shallow_copy(
-        self, extra_dataset_patterns: Patterns | None = None
-    ) -> KedroDataCatalog:
-        # TODO: remove when removing old catalog
-        """Returns a shallow copy of the current object.
-
-        Returns:
-            Copy of the current object.
-        """
-        if extra_dataset_patterns:
-            self._config_resolver.add_runtime_patterns(extra_dataset_patterns)
-        return self
 
     def exists(self, name: str) -> bool:
         """Checks whether registered dataset exists by calling its `exists()`
@@ -822,7 +671,58 @@ class KedroDataCatalog(CatalogProtocol):
 
         """
         try:
-            dataset = self._get_dataset(name)
+            dataset = self.get(name)
         except DatasetNotFoundError:
             return False
         return dataset.exists()
+
+    @staticmethod
+    def _validate_versions(
+        datasets: dict[str, AbstractDataset] | None,
+        load_versions: dict[str, str],
+        save_version: str | None,
+    ) -> tuple[dict[str, str], str | None]:
+        """Validates and synchronises dataset versions for loading and saving.
+
+        Ensures consistency of dataset versions across a catalog, particularly
+        for versioned datasets. It updates load versions and validates that all
+        save versions are consistent.
+
+        Args:
+            datasets: A dictionary mapping dataset names to their instances.
+                if None, no validation occurs.
+            load_versions: A mapping between dataset names and versions
+                to load.
+            save_version: Version string to be used for ``save`` operations
+                by all datasets with versioning enabled.
+
+        Returns:
+            Updated ``load_versions`` with load versions specified in the ``datasets``
+                and resolved ``save_version``.
+
+        Raises:
+            VersionAlreadyExistsError: If a dataset's save version conflicts with
+                the catalog's save version.
+        """
+        if not datasets:
+            return load_versions, save_version
+
+        cur_load_versions = load_versions.copy()
+        cur_save_version = save_version
+
+        for ds_name, ds in datasets.items():
+            cur_ds = ds._dataset if isinstance(ds, CachedDataset) else ds
+
+            if isinstance(cur_ds, AbstractVersionedDataset) and cur_ds._version:
+                if cur_ds._version.load:
+                    cur_load_versions[ds_name] = cur_ds._version.load
+                if cur_ds._version.save:
+                    cur_save_version = cur_save_version or cur_ds._version.save
+                    if cur_save_version != cur_ds._version.save:
+                        raise VersionAlreadyExistsError(
+                            f"Cannot add a dataset `{ds_name}` with `{cur_ds._version.save}` save version. "
+                            f"Save version set for the catalog is `{cur_save_version}`"
+                            f"All datasets in the catalog must have the same save version."
+                        )
+
+        return cur_load_versions, cur_save_version
